Refactored how React/DevTools log Timeline performance data
Until now, DEV and PROFILING builds of React recorded Timeline profiling data using the User Timing V3 API. This commit changes things so that React records this data by calling methods on the DevTools hook. (For now, DevTools still records that data using the User Timing V3 API, to match previous behavior.)

This commit is mostly moving around a lot of things:
* New methods have been added to the DevTools hook (in "backend/profilingHooks") for recording the Timeline perfromance events.
* Reconciler's "ReactFiberDevToolsHook" has been updated to call these new methods (when they're present).
* User Timing method calls in "SchedulingProfiler" have been moved to DevTools "backend/profilingHooks" (to match previous behavior, for now).
* The old reconciler tests, "SchedulingProfiler-test" and "SchedulingProfilerLabels-test", have been moved into DevTools "TimelineProfiler-test".
* Two new methods have been added to the injected renderer interface: "injectProfilingHooks" and "getLaneLabelMap"
